added common header file for audio helper functions
