+++ utilities/test_suite/HOST/Tensor_host_audio.cpp
@@ -20,137 +20,7 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 THE SOFTWARE.
 */
 
-#include "../rpp_test_suite_common.h"
-#include <half/half.hpp>
-#include <fstream>
-#include <iomanip>
-#include <vector>
-
-namespace fs = boost::filesystem;
-
-// Include this header file to use functions from libsndfile
-#include <sndfile.h>
-
-using namespace std;
-using half_float::half;
-
-typedef half Rpp16f;
-
-void verify_output(Rpp32f *dstPtr, RpptDescPtr dstDescPtr, RpptImagePatchPtr dstDims, string testCase, vector<string> audioNames, string dst)
-{
-    fstream refFile;
-    string refPath = get_current_dir_name();
-    string pattern = "HOST/build";
-    remove_substring(refPath, pattern);
-    refPath = refPath + "REFERENCE_OUTPUTS_AUDIO/";
-    int fileMatch = 0;
-    for (int batchCount = 0; batchCount < dstDescPtr->n; batchCount++)
-    {
-        string currentFileName = audioNames[batchCount];
-        size_t lastIndex = currentFileName.find_last_of(".");
-        currentFileName = currentFileName.substr(0, lastIndex);  // Remove extension from file name
-        string outFile = refPath + testCase + "/" + testCase + "_ref_" + currentFileName + ".txt";
-        refFile.open(outFile, ios::in);
-        if (!refFile.is_open())
-        {
-            cout << "\n Unable to open the file specified! Please check the path of the file given as input" << endl;
-            break;
-        }
-        int matchedIndices = 0;
-        Rpp32f refVal, outVal;
-        Rpp32f *dstPtrCurrent = dstPtr + batchCount * dstDescPtr->strides.nStride;
-        Rpp32f *dstPtrRow = dstPtrCurrent;
-        for (int i = 0; i < dstDims[batchCount].height; i++)
-        {
-            Rpp32f *dstPtrTemp = dstPtrRow;
-            for (int j = 0; j < dstDims[batchCount].width; j++)
-            {
-                refFile >> refVal;
-                outVal = dstPtrTemp[j];
-                bool invalidComparision = ((outVal == 0.0f) && (refVal != 0.0f));
-                if (!invalidComparision && abs(outVal - refVal) < 1e-20)
-                    matchedIndices += 1;
-            }
-            dstPtrRow += dstDescPtr->strides.hStride;
-        }
-        refFile.close();
-        if (matchedIndices == (dstDims[batchCount].width * dstDims[batchCount].height) && matchedIndices !=0)
-            fileMatch++;
-    }
-    std::string status = testCase + ": ";
-    cout << std::endl << "Results for Test case: " << testCase << std::endl;
-    if (fileMatch == dstDescPtr->n)
-    {
-        cout << "PASSED!" << std::endl;
-        status += "PASSED";
-    }
-    else
-    {
-        cout << "FAILED! " << fileMatch << "/" << dstDescPtr->n << " outputs are matching with reference outputs" << std::endl;
-        status += "FAILED";
-    }
-    std::string qaResultsPath = dst + "/QA_results.txt";
-    std:: ofstream qaResults(qaResultsPath, ios_base::app);
-    if (qaResults.is_open())
-    {
-        qaResults << status << std::endl;
-        qaResults.close();
-    }
-}
-
-void verify_non_silent_region_detection(float *detectedIndex, float *detectionLength, string testCase, int bs, vector<string> audioNames, string dst)
-{
-    fstream refFile;
-    string refPath = get_current_dir_name();
-    string pattern = "HOST/build";
-    remove_substring(refPath, pattern);
-    refPath = refPath + "REFERENCE_OUTPUTS_AUDIO/";
-    int fileMatch = 0;
-    for (int i = 0; i < bs; i++)
-    {
-        string currentFileName = audioNames[i];
-        size_t lastIndex = currentFileName.find_last_of(".");
-        currentFileName = currentFileName.substr(0, lastIndex);  // Remove extension from file name
-        string outFile = refPath + testCase + "/" + testCase + "_ref_" + currentFileName + ".txt";
-        refFile.open(outFile, ios::in);
-        if (!refFile.is_open())
-        {
-            cout << "\n Unable to open the file specified! Please check the path of the file given as input" << endl;
-            break;
-        }
-
-        Rpp32s refIndex, refLength;
-        Rpp32s outIndex, outLength;
-        refFile >> refIndex;
-        refFile >> refLength;
-        outIndex = detectedIndex[i];
-        outLength = detectionLength[i];
-
-        if ((outIndex == refIndex) && (outLength == refLength))
-            fileMatch += 1;
-        refFile.close();
-    }
-    std::string status = testCase + ": ";
-    cout << std::endl << "Results for Test case: " << testCase << std::endl;
-    if (fileMatch == bs)
-    {
-        cout << "PASSED!" << std::endl;
-        status += "PASSED";
-    }
-    else
-    {
-        cout << "FAILED! "<< fileMatch << "/" << bs << " outputs are matching with reference outputs" << std::endl;
-        status += "FAILED";
-    }
-
-    std::string qaResultsPath = dst + "/QA_results.txt";
-    std:: ofstream qaResults(qaResultsPath, ios_base::app);
-    if (qaResults.is_open())
-    {
-        qaResults << status << std::endl;
-        qaResults.close();
-    }
-}
+#include "../rpp_test_suite_audio.h"
 
 int main(int argc, char **argv)
 {
